requirements.txt
google-cloud-aiplatform>=1.36.0
google-cloud-storage>=2.8.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
vertexai>=0.0.1
requests>=2.31.0


